Remove tear_down_with_tasks from MountingSourceBase
The MountingSourceBase class will switch to mixin class.
